Move tx methods; add sig error (#1021)
