+++ src/ethereum/exceptions.py
@@ -39,3 +39,9 @@ class InvalidSenderError(InvalidTransaction):
     Thrown when a transaction originates from an account that cannot send
     transactions.
     """
+
+
+class InvalidSignatureError(InvalidTransaction):
+    """
+    Thrown when a transaction has an invalid signature.
+    """
