+++ src/ethereum/muir_glacier/transactions.py
@@ -8,8 +8,13 @@ from typing import Union
 
 from ethereum_types.bytes import Bytes, Bytes0
 from ethereum_types.frozen import slotted_freezable
-from ethereum_types.numeric import U256, Uint
+from ethereum_types.numeric import U64, U256, Uint
 
+from ethereum.crypto.elliptic_curve import SECP256K1N, secp256k1_recover
+from ethereum.crypto.hash import Hash32, keccak256
+from ethereum.exceptions import InvalidSignatureError
+
+from .. import rlp
 from .fork_types import Address
 
 TX_BASE_COST = 21000
@@ -34,3 +39,177 @@ class Transaction:
     v: U256
     r: U256
     s: U256
+
+
+def validate_transaction(tx: Transaction) -> bool:
+    """
+    Verifies a transaction.
+
+    The gas in a transaction gets used to pay for the intrinsic cost of
+    operations, therefore if there is insufficient gas then it would not
+    be possible to execute a transaction and it will be declared invalid.
+
+    Additionally, the nonce of a transaction must not equal or exceed the
+    limit defined in `EIP-2681 <https://eips.ethereum.org/EIPS/eip-2681>`_.
+    In practice, defining the limit as ``2**64-1`` has no impact because
+    sending ``2**64-1`` transactions is improbable. It's not strictly
+    impossible though, ``2**64-1`` transactions is the entire capacity of the
+    Ethereum blockchain at 2022 gas limits for a little over 22 years.
+
+    Parameters
+    ----------
+    tx :
+        Transaction to validate.
+
+    Returns
+    -------
+    verified : `bool`
+        True if the transaction can be executed, or False otherwise.
+    """
+    if calculate_intrinsic_cost(tx) > Uint(tx.gas):
+        return False
+    if tx.nonce >= U256(U64.MAX_VALUE):
+        return False
+    return True
+
+
+def calculate_intrinsic_cost(tx: Transaction) -> Uint:
+    """
+    Calculates the gas that is charged before execution is started.
+
+    The intrinsic cost of the transaction is charged before execution has
+    begun. Functions/operations in the EVM cost money to execute so this
+    intrinsic cost is for the operations that need to be paid for as part of
+    the transaction. Data transfer, for example, is part of this intrinsic
+    cost. It costs ether to send data over the wire and that ether is
+    accounted for in the intrinsic cost calculated in this function. This
+    intrinsic cost must be calculated and paid for before execution in order
+    for all operations to be implemented.
+
+    Parameters
+    ----------
+    tx :
+        Transaction to compute the intrinsic cost of.
+
+    Returns
+    -------
+    verified : `ethereum.base_types.Uint`
+        The intrinsic cost of the transaction.
+    """
+    data_cost = 0
+
+    for byte in tx.data:
+        if byte == 0:
+            data_cost += TX_DATA_COST_PER_ZERO
+        else:
+            data_cost += TX_DATA_COST_PER_NON_ZERO
+
+    if tx.to == Bytes0(b""):
+        create_cost = TX_CREATE_COST
+    else:
+        create_cost = 0
+
+    return Uint(TX_BASE_COST + data_cost + create_cost)
+
+
+def recover_sender(chain_id: U64, tx: Transaction) -> Address:
+    """
+    Extracts the sender address from a transaction.
+
+    The v, r, and s values are the three parts that make up the signature
+    of a transaction. In order to recover the sender of a transaction the two
+    components needed are the signature (``v``, ``r``, and ``s``) and the
+    signing hash of the transaction. The sender's public key can be obtained
+    with these two values and therefore the sender address can be retrieved.
+
+    Parameters
+    ----------
+    tx :
+        Transaction of interest.
+    chain_id :
+        ID of the executing chain.
+
+    Returns
+    -------
+    sender : `ethereum.fork_types.Address`
+        The address of the account that signed the transaction.
+    """
+    v, r, s = tx.v, tx.r, tx.s
+    if U256(0) >= r or r >= SECP256K1N:
+        raise InvalidSignatureError("bad r")
+    if U256(0) >= s or s > SECP256K1N // U256(2):
+        raise InvalidSignatureError("bad s")
+
+    if v == 27 or v == 28:
+        public_key = secp256k1_recover(
+            r, s, v - U256(27), signing_hash_pre155(tx)
+        )
+    else:
+        chain_id_x2 = U256(chain_id) * U256(2)
+        if v != U256(35) + chain_id_x2 and v != U256(36) + chain_id_x2:
+            raise InvalidSignatureError("bad v")
+        public_key = secp256k1_recover(
+            r, s, v - U256(35) - chain_id_x2, signing_hash_155(tx, chain_id)
+        )
+    return Address(keccak256(public_key)[12:32])
+
+
+def signing_hash_pre155(tx: Transaction) -> Hash32:
+    """
+    Compute the hash of a transaction used in a legacy (pre EIP 155) signature.
+
+    Parameters
+    ----------
+    tx :
+        Transaction of interest.
+
+    Returns
+    -------
+    hash : `ethereum.crypto.hash.Hash32`
+        Hash of the transaction.
+    """
+    return keccak256(
+        rlp.encode(
+            (
+                tx.nonce,
+                tx.gas_price,
+                tx.gas,
+                tx.to,
+                tx.value,
+                tx.data,
+            )
+        )
+    )
+
+
+def signing_hash_155(tx: Transaction, chain_id: U64) -> Hash32:
+    """
+    Compute the hash of a transaction used in a EIP 155 signature.
+
+    Parameters
+    ----------
+    tx :
+        Transaction of interest.
+    chain_id :
+        The id of the current chain.
+
+    Returns
+    -------
+    hash : `ethereum.crypto.hash.Hash32`
+        Hash of the transaction.
+    """
+    return keccak256(
+        rlp.encode(
+            (
+                tx.nonce,
+                tx.gas_price,
+                tx.gas,
+                tx.to,
+                tx.value,
+                tx.data,
+                chain_id,
+                Uint(0),
+                Uint(0),
+            )
+        )
+    )
