+++ src/ethereum_spec_tools/evm_tools/loaders/fork_loader.py
@@ -80,32 +80,32 @@ class ForkLoad:
     @property
     def signing_hash(self) -> Any:
         """signing_hash function of the fork"""
-        return self._module("fork").signing_hash
+        return self._module("transactions").signing_hash
 
     @property
     def signing_hash_pre155(self) -> Any:
         """signing_hash_pre155 function of the fork"""
-        return self._module("fork").signing_hash_pre155
+        return self._module("transactions").signing_hash_pre155
 
     @property
     def signing_hash_155(self) -> Any:
         """signing_hash_155 function of the fork"""
-        return self._module("fork").signing_hash_155
+        return self._module("transactions").signing_hash_155
 
     @property
     def signing_hash_2930(self) -> Any:
         """signing_hash_2930 function of the fork"""
-        return self._module("fork").signing_hash_2930
+        return self._module("transactions").signing_hash_2930
 
     @property
     def signing_hash_1559(self) -> Any:
         """signing_hash_1559 function of the fork"""
-        return self._module("fork").signing_hash_1559
+        return self._module("transactions").signing_hash_1559
 
     @property
     def signing_hash_4844(self) -> Any:
         """signing_hash_4844 function of the fork"""
-        return self._module("fork").signing_hash_4844
+        return self._module("transactions").signing_hash_4844
 
     @property
     def check_transaction(self) -> Any:
